move routines for xml based filtering from 'distributionUtiils.py' to 'filters.py', add more unit tests
